fi: salary int to earnedincome double

+++ app/models/fi/PerDiemPayment.scala
@@ -2,7 +2,7 @@ package models.fi
 
 import play.api.libs.json.{Json, JsObject}
 
-class PerDiemPayment(taxableSalary: Int) {
+class PerDiemPayment(taxableSalary: Double) {
   private val perDiemPaymentsTyelPercent = 0.0078
 
   def getSum: Double = {
